+++ src/__tests__/validate.test.ts
@@ -52,7 +52,7 @@ const invalidData: any[] = [
 ];
 
 describe('validate() middleware', () => {
-    ['delete', 'options', 'patch', 'put', 'post', 'trace'].forEach(method => {
+    ['patch', 'put', 'post'].forEach(method => {
         const methodName = method.toUpperCase();
 
         it.each(validData)(`should return 200 when do a ${methodName} request and send valid JSON data`, async (vd) => {